requirements.txt
python-barcode
qrcode[pil]
